+++ src/tools/query.py
@@ -6,7 +6,6 @@ This module contains tools for getting the schema for a collection and running S
 
 import logging
 import re
-from collections import Counter
 from typing import Any
 
 from lark_sqlpp import modifies_data, modifies_structure, parse_sqlpp
@@ -15,6 +14,10 @@ from mcp.server.fastmcp import Context
 from utils.connection import connect_to_bucket
 from utils.constants import MCP_SERVER_NAME
 from utils.context import get_cluster_connection
+from utils.query_utils import (
+    evaluate_query_plan,
+    extract_plan_from_explain_results,
+)
 
 logger = logging.getLogger(f"{MCP_SERVER_NAME}.tools.query")
 
@@ -108,136 +111,6 @@ def run_sql_plus_plus_query(
         raise
 
 
-def _extract_plan_from_explain_results(
-    explain_results: list[dict[str, Any]],
-) -> dict[str, Any] | None:
-    """Extract the plan object from an EXPLAIN query response."""
-    if not explain_results:
-        return None
-
-    first_row = explain_results[0]
-    if not isinstance(first_row, dict):
-        return None
-
-    plan = first_row.get("plan")
-    if isinstance(plan, dict):
-        return plan
-
-    if "#operator" in first_row or "~children" in first_row:
-        return first_row
-
-    for value in first_row.values():
-        if isinstance(value, dict) and (
-            "#operator" in value or "~children" in value or "plan" in value
-        ):
-            nested_plan = value.get("plan") if "plan" in value else value
-            if isinstance(nested_plan, dict):
-                return nested_plan
-
-    return None
-
-
-def _walk_plan(
-    node: Any,
-    operators: list[str],
-    indexes_used: set[str],
-    keyspaces: set[str],
-) -> None:
-    """Recursively walk an EXPLAIN plan node and collect details."""
-    if isinstance(node, dict):
-        operator = node.get("#operator")
-        if isinstance(operator, str):
-            operators.append(operator)
-
-        index_name = node.get("index")
-        if isinstance(index_name, str) and index_name:
-            indexes_used.add(index_name)
-
-        keyspace_name = node.get("keyspace")
-        if isinstance(keyspace_name, str) and keyspace_name:
-            keyspaces.add(keyspace_name)
-
-        for value in node.values():
-            _walk_plan(value, operators, indexes_used, keyspaces)
-    elif isinstance(node, list):
-        for item in node:
-            _walk_plan(item, operators, indexes_used, keyspaces)
-
-
-def evaluate_query_plan(plan: dict[str, Any] | None) -> dict[str, Any]:
-    """Evaluate an EXPLAIN plan and return optimization findings."""
-    if not plan:
-        return {
-            "summary": "No query plan found in EXPLAIN output.",
-            "operators": [],
-            "operator_counts": {},
-            "indexes_used": [],
-            "keyspaces": [],
-            "findings": [],
-        }
-
-    operators: list[str] = []
-    indexes_used: set[str] = set()
-    keyspaces: set[str] = set()
-
-    _walk_plan(plan, operators, indexes_used, keyspaces)
-
-    operator_counts = dict(Counter(operators))
-    findings: list[dict[str, str]] = []
-
-    has_primary_scan = any(op.startswith("PrimaryScan") for op in operators)
-    has_secondary_index_scan = any(op.startswith("IndexScan") for op in operators)
-    has_fetch = "Fetch" in operator_counts
-
-    if has_primary_scan:
-        findings.append(
-            {
-                "severity": "warning",
-                "issue": "primary_index_scan",
-                "message": (
-                    "Primary index scan detected. Consider creating a targeted "
-                    "secondary index for better selectivity."
-                ),
-            }
-        )
-
-    if has_fetch and has_secondary_index_scan:
-        findings.append(
-            {
-                "severity": "warning",
-                "issue": "non_covering_index",
-                "message": (
-                    "Fetch operator detected after secondary index scan. "
-                    "A covering index may reduce document fetches."
-                ),
-            }
-        )
-
-    if not findings:
-        findings.append(
-            {
-                "severity": "info",
-                "issue": "no_obvious_issues",
-                "message": ("No common anti-patterns detected from the query plan."),
-            }
-        )
-
-    summary = (
-        "Plan has optimization opportunities."
-        if any(f["severity"] == "warning" for f in findings)
-        else "Plan looks healthy for common query-plan checks."
-    )
-
-    return {
-        "summary": summary,
-        "operators": sorted(operator_counts),
-        "operator_counts": operator_counts,
-        "indexes_used": sorted(indexes_used),
-        "keyspaces": sorted(keyspaces),
-        "findings": findings,
-    }
-
-
 def explain_sql_plus_plus_query(
     ctx: Context,
     bucket_name: str,
@@ -246,7 +119,8 @@ def explain_sql_plus_plus_query(
 ) -> dict[str, Any]:
     """Generate and evaluate an EXPLAIN plan for a SQL++ query. It provides information about the execution plan for the query.
 
-    This tool is stateless: callers must provide query and scope context explicitly.
+    The EXPLAIN statement is run in the specified scope in the specified bucket.
+    It returns the raw EXPLAIN response along with an extracted plan and plan evaluation.
     """
     normalized_query = query.strip()
     if not normalized_query:
@@ -265,17 +139,13 @@ def explain_sql_plus_plus_query(
         explain_statement,
     )
 
-    plan = _extract_plan_from_explain_results(explain_results)
+    plan = extract_plan_from_explain_results(explain_results)
     plan_evaluation = evaluate_query_plan(plan)
 
     return {
         "query": query,
         "explain_statement": explain_statement,
-        "query_context": {
-            "bucket_name": bucket_name,
-            "scope_name": scope_name,
-            "source": "explicit",
-        },
+        "query_context": {"bucket_name": bucket_name, "scope_name": scope_name},
         "plan": plan,
         "plan_evaluation": plan_evaluation,
         "results": explain_results,
